refactor(knowledge-store): rewire the tasks, janitor, and seeder onto the facade
The janitor, the seeder, the index tasks, and the fleet runner all reach
the store through service.py, closing out the old store/write_lock imports.
