+++ surfsense_backend/scripts/migrate_knowledge_store.py
@@ -19,7 +19,7 @@ from sqlalchemy import select, update
 
 from app.db import Workspace, async_session_maker
 from app.knowledge_store.migrate import migrate_workspace
-from app.knowledge_store.store import KnowledgeStore
+from app.knowledge_store import KnowledgeStore
 
 
 async def _workspace_ids(only: list[int]) -> list[int]:
